Rebuilt to scroll by view window height
Pinterest unloads pins that aren't currently on the screen, so scrolling all the way down to the bottom and then scraping for links doesn't work. Instead this script will scroll by viewport heights and add new links to a set, then download at the end.

Currently it gets profile pics as well